+++ packages/editor/src/components/commands/index.js
@@ -545,7 +545,7 @@ const getPageContentFocusCommands = () =>
 		if ( !! goBack ) {
 			commands.push( {
 				name: 'core/switch-to-previous-entity',
-				label: __( 'Go back' ),
+				label: __( 'Back' ),
 				category: 'view',
 				callback: ( { close } ) => {
 					goBack();